class-32/demo/requirements.txt
asgiref==3.4.1; python_version >= "3.8"
django==4.0; python_version >= "3.8"
djangorestframework==3.12.4; python_version >= "3.5"
pytz==2021.3
sqlparse==0.4.2; python_version >= "3.8"
tzdata==2021.5; sys_platform == "win32" and python_version >= "3.8"
# poetry export -f requirements.txt -o requirements.txt --without-hashes